requirements.txt
duckdb==1.2.1
pyarrow==16.1.0
flask==3.1.0
gunicorn==23.0.0
google-cloud-storage==2.19.0
google-cloud-bigquery==3.29.0
google-cloud-pubsub==2.28.0
fsspec==2025.3.0
gcsfs==2025.3.0
pandas==2.2.3
